fixup! Patch upgrade

+++ test/utils/utils.sh
@@ -90,9 +90,10 @@ jenkins_updates () {
 	# Update UCS@school instances always to latest patchlevel version
 	[ -z "$target" ] && target="$(echo "${JOB_NAME:-}"|sed -rne 's,^UCSschool-([0-9]+\.[0-9]+)/.*,\1-99,p')"
 
-	if [ $(ucr get version/version) = "4.4" ]; then
+	if [ $(ucr get version/version) = "4.4" ] && [ -e "/root/0001-Upgrade.patch" ]; then
 		apt-get install -y patch
 		patch -p 1 -d / -i /root/0001-Upgrade.patch || exit $?
+		rm -f /root/0001-Upgrade.patch
 	fi
 
 	test -n "$TARGET_VERSION" && target="$TARGET_VERSION"
